Fix many dumb mistakes
